mce: capture status of per-node operations
in mce_for_each_node.  This replicates the status-handling logic of
mce_for_each_engine, and ensures that we exit with the largest non-zero
status returned by any of the node-specific function calls -- which
means that "mce <options> offline" will exit with a nonzero status if
one or more engines could not be successfully removed from the cluster.